% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement of DNA sequences}
\usage{
reverse_complement(seqs)
}
\arguments{
\item{seqs}{Character vector of DNA sequences containing only A, C, G, T
or N (any case).}
}
\value{
Character vector of upper-case reverse complements, named like \code{seqs}.
}
\description{
Reverse complement of DNA sequences
}
\examples{
reverse_complement("ATGCCC")
reverse_complement(c(fwd = "GGATCCNA"))
}
